refactored HTTP protocol and updated README
